Move CheckoutSession to the Checkout namespace and rename to Session
